Добавил модалку при удалении профиля, исправил недочеты стилей
