+++ react-chat/src/components/header/HeaderChat.jsx
@@ -1,5 +1,7 @@
-import React from 'react'
+import React, { useEffect } from 'react'
+
 import { Link } from 'react-router-dom'
+import { useNavigate } from 'react-router-dom'
 
 import backButtonLogo from '/backButton.svg'
 import menuButtonLogo from '/menuButton.svg'
@@ -9,31 +11,41 @@ import searchButtonLogo from '/searchButton.svg'
 import './HeaderChat.css'
 
 export const HeaderChat = ({ user }) => {
+    const navigate = useNavigate();
+
+    useEffect(() => {
+      if (!user) {
+        navigate('/');
+      }
+    }, [])
+
     return (
-        <header className='chat-header'>
+        user && (
+            <header className='chat-header'>
 
-        <Link to='/' className='back-button'>
-          <img src={backButtonLogo}/>
-        </Link>
+            <Link to='/' className='back-button'>
+              <img src={backButtonLogo}/>
+            </Link>
 
-        <div className='profile-link'>
-          <img src={user.avatar || profileLinkLogo}/>
-        </div>
+            <div className='profile-link'>
+              <img src={user.avatar || profileLinkLogo}/>
+            </div>
 
-        <div className='user-info'>
-          <h3>{user.first_name + ' ' +  user.last_name || 'User'}</h3>
-          <p>last seen 2 hours ago</p>
-        </div>
+            <div className='user-info'>
+              <h3>{user.first_name + ' ' +  user.last_name || 'User'}</h3>
+              <p>last seen 2 hours ago</p>
+            </div>
 
-        <div className='right-icons'>
-          <button>
-            <img src={searchButtonLogo}/>
-          </button>
-          <button>
-            <img src={menuButtonLogo}/>
-          </button>
-        </div>
+            <div className='right-icons'>
+              <button>
+                <img src={searchButtonLogo}/>
+              </button>
+              <button>
+                <img src={menuButtonLogo}/>
+              </button>
+            </div>
 
-      </header>
+          </header>
+        )
     );
 };
