+++ react-chat/src/components/message-box/index.css
@@ -1,0 +1,26 @@
+.message-box-container {
+    display: flex;
+    flex-direction: column;
+    position: fixed;
+    top: 50%;
+    left: 50%;
+    transform: translate(-50%, -50%);
+    background-color: #fff;
+    padding: 20px;
+    border-radius: 8px;
+    box-shadow: 0 4px 6px rgba(0, 0, 0, 0.1);
+    z-index: 100000000000000;
+}
+
+.confirm-buttons {
+    display: flex;
+    flex-direction: row;
+    justify-content: space-between;
+    margin-top: 15px;
+}
+
+.message-box-text {
+    font-size: large;
+    font-weight: bold;
+    color: red;
+}
