+++ react-chat/src/components/send_photos/index.css
@@ -4,6 +4,7 @@
     justify-content: center;
     gap: 10px;
     padding: 20px;
+    z-index: 1000000;
 }
 
 .image-item {
